Issue QD-2400
Added the universal chunk for all local runs.
Extended the local run section with the Qodana CLI option.
Moved all content from the linter-related and the Getting Started page to the universal chunk.
Added the chunk references to the linter-related and the Getting Started pages.
Modified the project.ihp file to build this version as a separate instance.

This needs to be built and reviewed. Work in Progress.